skillsyncAPI/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
google-generativeai
markdown2
